+++ client/src/components/Search.tsx
@@ -16,6 +16,12 @@ const useStyles = makeStyles((theme: Theme) => ({
     boxShadow: 'none',
     margin: '50px auto',
     width: '70%',
+    height: '100px',
+    position: 'absolute',
+    top: '3rem',
+    bottom: 0,
+    left: 0,
+    right: 0,
   },
   searchInput: {
     display: 'flex',
