+++ client/src/components/Users/Auth/SignUpCitizen/interests.ts
@@ -1,0 +1,19 @@
+export const interests: string[] = [
+  'Animal Care & Services',
+  'Poverty',
+  'Housing & Homeless',
+  'Youth & Children',
+  'Disaster Relief',
+  'Health Care & Welness',
+  'Environment & Sustainability',
+  'Sports & Recreation',
+  'Seniors',
+  'Religion, Faith & Spirituality',
+  'Civic Engagement',
+  'LGTBQIA+',
+  'Civil Rights & Advocacy',
+  'Military & Veterans',
+  'Social Justice',
+  'Education & Literacy',
+  'Arts & Culture',
+];
