+++ client/src/theme/typography.ts
@@ -11,6 +11,13 @@ const typography = {
     fontSize: '2.8rem',
     fontWeight: 'bold',
   },
+  h2: {
+    fontSize: '3.625rem',
+    fontWeight: 400,
+    '@media (max-width:600px)': {
+      fontSize: '2.5rem',
+    },
+  },
   h3: {
     fontSize: '2.2rem',
     fontWeight: 'bold',
