+++ client/src/views/TermsOfService.tsx
@@ -1,7 +1,17 @@
 import * as React from 'react';
+import { Theme } from '@mui/material/styles';
+import makeStyles from '@mui/styles/makeStyles';
 
+const useStyles = makeStyles((theme: Theme) => {
+  return {
+    main: {
+      height: '100vh',
+    },
+  };
+});
 function TermsOfService() {
-  return <div>TermsOfService</div>;
+  const classes = useStyles();
+  return <div className={classes.main}>TermsOfService</div>;
 }
 
 export default TermsOfService;
